# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog API - premium gym workout tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MuscleGroup = (typeof MuscleGroup)[keyof typeof MuscleGroup];

export const MuscleGroup = {
  chest: "chest",
  back: "back",
  legs: "legs",
  shoulders: "shoulders",
  arms: "arms",
  core: "core",
  full_body: "full_body",
} as const;

export type EquipmentLevel =
  (typeof EquipmentLevel)[keyof typeof EquipmentLevel];

export const EquipmentLevel = {
  full_gym: "full_gym",
  machines_only: "machines_only",
  dumbbells_only: "dumbbells_only",
  home: "home",
} as const;

export type Weekday = (typeof Weekday)[keyof typeof Weekday];

export const Weekday = {
  monday: "monday",
  tuesday: "tuesday",
  wednesday: "wednesday",
  thursday: "thursday",
  friday: "friday",
  saturday: "saturday",
  sunday: "sunday",
} as const;

export type Difficulty = (typeof Difficulty)[keyof typeof Difficulty];

export const Difficulty = {
  easy: "easy",
  moderate: "moderate",
  hard: "hard",
  failure: "failure",
} as const;

export type Units = (typeof Units)[keyof typeof Units];

export const Units = {
  kg: "kg",
  lb: "lb",
} as const;

export interface Profile {
  id: string;
  displayName: string;
  equipment: EquipmentLevel;
  units: Units;
  createdAt: string;
}

export interface UpdateProfileBody {
  displayName?: string;
  equipment?: EquipmentLevel;
  units?: Units;
}

export interface Exercise {
  id: string;
  name: string;
  muscleGroup: MuscleGroup;
  equipment: EquipmentLevel;
  isCustom: boolean;
  description?: string;
  formCues: string[];
}

export interface CreateExerciseBody {
  name: string;
  muscleGroup: MuscleGroup;
  equipment: EquipmentLevel;
  description?: string;
  formCues?: string[];
}

export interface PlanDay {
  day: Weekday;
  label: string;
  isRest?: boolean;
  exercises: Exercise[];
}

export interface UpdatePlanDayBody {
  label: string;
  isRest?: boolean;
  exerciseIds: string[];
}

export interface SetEntry {
  id: string;
  sessionExerciseId: string;
  setNumber: number;
  weight: number;
  reps: number;
  difficulty: Difficulty;
  notes?: string;
  createdAt: string;
}

export interface SessionExercise {
  id: string;
  exercise: Exercise;
  sets: SetEntry[];
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface Session {
  id: string;
  day: Weekday;
  label: string;
  status: SessionStatus;
  startedAt: string;
  completedAt?: string;
  durationSec?: number;
  totalVolume?: number;
  estimatedCalories?: number;
  exercises: SessionExercise[];
}

export type SessionSummaryStatus =
  (typeof SessionSummaryStatus)[keyof typeof SessionSummaryStatus];

export const SessionSummaryStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface SessionSummary {
  id: string;
  day: Weekday;
  label: string;
  status: SessionSummaryStatus;
  startedAt: string;
  completedAt?: string;
  durationSec?: number;
  totalVolume?: number;
  estimatedCalories?: number;
  totalSets: number;
}

export interface StartSessionBody {
  day: Weekday;
  label: string;
  exerciseIds: string[];
}

export interface CompleteSessionBody {
  durationSec?: number;
}

export interface AddSetBody {
  exerciseId: string;
  weight: number;
  reps: number;
  difficulty: Difficulty;
  notes?: string;
}

export interface PersonalRecord {
  exerciseId: string;
  exerciseName: string;
  weight: number;
  reps: number;
  estimatedOneRm?: number;
  achievedAt: string;
}

export interface SessionCompletion {
  session: Session;
  newPRs: PersonalRecord[];
  currentStreak: number;
}

export interface StatsSummary {
  currentStreak: number;
  longestStreak: number;
  workoutsThisWeek: number;
  workoutsAllTime: number;
  totalVolumeAllTime: number;
  /** 0-100 based on last 4 weeks */
  consistencyScore: number;
  lastWorkoutDay: string;
  skippedLegDay?: boolean;
}

export interface VolumePoint {
  weekStart: string;
  volume: number;
}

export interface FrequencyPoint {
  weekStart: string;
  count: number;
}

export interface StrengthPoint {
  date: string;
  topWeight: number;
  topReps: number;
  estimatedOneRm: number;
}

export interface BodyweightEntry {
  id: string;
  weight: number;
  recordedAt: string;
}

export interface AddBodyweightBody {
  weight: number;
}

export interface WaterToday {
  date: string;
  cups: number;
  goalCups: number;
}

export type ListExercisesParams = {
  muscleGroup?: MuscleGroup;
  equipment?: EquipmentLevel;
};

export type ListSessionsParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};