# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IronLog API - premium gym workout tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises in catalog
      parameters:
        - in: query
          name: muscleGroup
          required: false
          schema:
            $ref: "#/components/schemas/MuscleGroup"
        - in: query
          name: equipment
          required: false
          schema:
            $ref: "#/components/schemas/EquipmentLevel"
      responses:
        "200":
          description: Exercise list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Add a custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get an exercise with form guide
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exercise detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
...
  /plan/{day}:
    put:
      operationId: updatePlanDay
      tags: [plan]
      summary: Update a day's label and exercises
      parameters:
        - in: path
          name: day
          required: true
          schema:
            $ref: "#/components/schemas/Weekday"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanDayBody"
      responses:
        "200":
          description: Updated day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanDay"
...
components:
  schemas:
...
    MuscleGroup:
      type: string
      enum: [chest, back, legs, shoulders, arms, core, full_body]
    EquipmentLevel:
      type: string
      enum: [full_gym, machines_only, dumbbells_only, home]
    Weekday:
      type: string
      enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
...
    Exercise:
      type: object
      required: [id, name, muscleGroup, equipment, isCustom, formCues]
      properties:
        id:
          type: string
        name:
          type: string
        muscleGroup:
          $ref: "#/components/schemas/MuscleGroup"
        equipment:
          $ref: "#/components/schemas/EquipmentLevel"
        isCustom:
          type: boolean
        description:
          type: string
        formCues:
          type: array
          items:
            type: string
    CreateExerciseBody:
      type: object
      required: [name, muscleGroup, equipment]
      properties:
        name:
          type: string
        muscleGroup:
          $ref: "#/components/schemas/MuscleGroup"
        equipment:
          $ref: "#/components/schemas/EquipmentLevel"
        description:
          type: string
        formCues:
          type: array
          items:
            type: string
    PlanDay:
      type: object
      required: [day, label, exercises]
      properties:
        day:
          $ref: "#/components/schemas/Weekday"
        label:
          type: string
        isRest:
          type: boolean
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
    UpdatePlanDayBody:
      type: object
      required: [label, exerciseIds]
      properties:
        label:
          type: string
        isRest:
          type: boolean
        exerciseIds:
          type: array
          items:
            type: string
...