# lib/api-zod/src/generated/types/createExerciseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog API - premium gym workout tracker
 * OpenAPI spec version: 0.1.0
 */
import type { EquipmentLevel } from "./equipmentLevel";
import type { MuscleGroup } from "./muscleGroup";

export interface CreateExerciseBody {
  name: string;
  muscleGroup: MuscleGroup;
  equipment: EquipmentLevel;
  description?: string;
  formCues?: string[];
}