# lib/api-zod/src/generated/types/updatePlanDayBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog API - premium gym workout tracker
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePlanDayBody {
  label: string;
  isRest?: boolean;
  exerciseIds: string[];
}